# api_server/swagger.yaml
openapi: 3.0.0
info:
  title: Records API
  version: 1.0.0
  description: API for managing incident records
servers:
  - url: 'http://localhost:3000'
paths:
  /api/v1/records:
    get:
      summary: Get all records
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      summary: Create a new record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timestamp:
                  type: string
                Author:
                  type: string
                AuthorID:
                  type: string
                category:
                  type: string
                summary:
                  type: string
                system:
                  type: string
                Affected Objects:
                  type: string
                changed Values:
                  type: string
                IP address:
                  type: string
                Method:
                  type: string
                Node:
                  type: string
                extra Attributes:
                  type: object
                  properties:
                    Severity:
                      type: string
                    Tags:
                      type: array
                      items:
                        type: string
      responses:
        '201':
          description: Record created successfully
  /api/v1/records/{id}:
    get:
      summary: Get a record by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Record found
        '404':
          description: Record not found
    put:
      summary: Update a record by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Record updated successfully
        '404':
          description: Record not found
    delete:
      summary: Delete a record by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Record deleted successfully
        '404':
          description: Record not found